+++ src/boot/hk-components.js
@@ -25,6 +25,7 @@ import HkActionRollsTable from "../components/hk-components/hk-action-rolls/hk-a
 import HkActionRollScaling from "../components/hk-components/hk-action-rolls/hk-action-roll-scaling";
 import HkPane from "../components/hk-components/hk-pane";
 import HkFilter from "../components/hk-components/hk-filter";
+import HkTransformSelect from "../components/hk-components/hk-transform-select";
 
 export default async ({ Vue }) => {
 	Vue.component("hk-input", HkInput);
@@ -54,4 +55,5 @@ export default async ({ Vue }) => {
 	Vue.component("hk-action-roll-scaling", HkActionRollScaling);
 	Vue.component("hk-pane", HkPane);
 	Vue.component("hk-filter", HkFilter);
+	Vue.component("hk-transform-select", HkTransformSelect);
 };
